requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
flask-security-too==5.5.2
passlib[argon2]
Flask-WTF
gunicorn
gevent
python-dotenv
pyyaml
minio
setuptools
requests
Authlib
pytz
websockets==13.0
markdown
Pillow
gevent
redis
discord.py
babel
